+++ src/terra-dev-site/snt/secondary-scroll-test/multiple-files/example1.snt.jsx
@@ -1,0 +1,33 @@
+import React from 'react';
+import classNames from 'classnames/bind';
+import { injectIntl, intlShape } from 'react-intl';
+
+import styles from '../../Themed.module.scss';
+
+const cx = classNames.bind(styles);
+
+const propTypes = {
+  /**
+   * Internationalization object with translation APIs. Provided by `injectIntl`.
+   */
+  intl: intlShape.isRequired,
+};
+
+const Themed = ({ intl }) => (
+  <div className={cx('themed')}>
+    <h1>
+      Example File 1
+    </h1>
+    <div className={cx('themed-block')} />
+    <h1>
+      Translated block below
+    </h1>
+    <h1>
+      {intl.formatMessage({ id: 'Terra.devSite.themed.help' })}
+    </h1>
+  </div>
+);
+
+Themed.propTypes = propTypes;
+
+export default injectIntl(Themed);