+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Secondary navigation now scrolls to the current selected item
 
 6.7.0 - (November 1, 2019)
 ----------